fix View Order bug

+++ src/main/java/com/example/food/pojos/OrderResponse.java
@@ -32,7 +32,7 @@ public class OrderResponse extends BaseResponse {
     private Address address;
     private PaymentMethod paymentMethod;
     private BigDecimal deliveryFee;
-    private Double discount;
+    private BigDecimal discount;
     private DeliveryMethod deliveryMethod;
     private BigDecimal totalOrderPrice;
 }
